refactor tenant details dialog
- remove tenant context
- add react hook form
- create detailsDialog
- add delete and details icons in table rows
- add dialog for deleting tenants
