+++ src/components/tables/dataTable/dataTableToolbar.tsx
@@ -4,7 +4,6 @@ import Typography from "@mui/material/Typography";
 import Tooltip from "@mui/material/Tooltip";
 import IconButton from "@mui/material/IconButton";
 import DeleteIcon from "@mui/icons-material/Delete";
-import FilterListIcon from "@mui/icons-material/FilterList";
 import {Button} from "@mui/material";
 
 interface DataTableToolbarProps {
@@ -46,16 +45,10 @@ export default function DataTableToolbar({title, numSelected, deleteHandler, onC
                     {title}
                 </Typography>
             )}
-            {numSelected > 0 ? (
+            {numSelected > 0 && (
                 <Tooltip title="Delete">
                     <IconButton onClick={deleteHandler}>
                         <DeleteIcon/>
-                    </IconButton>
-                </Tooltip>
-            ) : (
-                <Tooltip title="Filter list">
-                    <IconButton>
-                        <FilterListIcon/>
                     </IconButton>
                 </Tooltip>
             )}
