+++ src/main.tsx
@@ -4,16 +4,13 @@ import App from './app.tsx'
 import './index.css'
 import {ApiAuthProvider} from "./context/apiAuthContext.tsx";
 import {ThemeModeProvider} from "./context/themeModeContext.tsx";
-import {TenantsProvider} from "./context/tenantsContext.tsx";
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
         <ApiAuthProvider>
-            <TenantsProvider>
-                <ThemeModeProvider>
-                    <App/>
-                </ThemeModeProvider>
-            </TenantsProvider>
+            <ThemeModeProvider>
+                <App/>
+            </ThemeModeProvider>
         </ApiAuthProvider>
     </React.StrictMode>,
 )
