+++ src/pages/tenantsPage.tsx
@@ -5,58 +5,41 @@ import {getTenants} from "../lib/api/getTenants.ts";
 import Typography from "@mui/material/Typography";
 import Box from "@mui/material/Box";
 import {useState} from "react";
-import TenantDetailsDialog from "../components/dialogs/TenantDetailsDialog";
-import {Tenant} from "../types/api";
-import DataTable from "../components/tables/dataTable.tsx";
-import {deleteTenant} from "../lib/api/deleteTenant.ts";
-import {HeadCell} from "../types/table";
+import {DeleteFnParams, Tenant, ValidationError} from "../types/api";
+import {AxiosResponse} from "axios";
+import TenantsTable from "../components/tables/tenantsTable.tsx";
+import TenantDetailsDialog from "../components/dialogs/tenantDetailsDialog.tsx";
 
 export default function TenantsPage() {
 
     const {apiPrefix, accessToken} = useApiAuthContext()
+
     const [tenants, setTenants] = useState<Tenant[]>([])
     const [currentTenant, setCurrentTenant] = useState<Tenant | null>(null)
     const [openDetails, setOpenDetails] = useState(false)
 
-    const headCells: HeadCell<Tenant>[] = [
-        {
-            id: "id",
-            numeric: false,
-            label: "id",
-            disablePadding: false,
-        },
-        {
-            id: "name",
-            numeric: false,
-            label: "name",
-            disablePadding: false,
-        },
-        {
-            id: "email",
-            numeric: false,
-            label: "email",
-            disablePadding: false,
-        },
-        {
-            id: "created",
-            numeric: false,
-            label: "created at",
-            disablePadding: false,
-        },
-        {
-            id: "expire",
-            numeric: false,
-            label: "expire at",
-            disablePadding: false,
-        },
-    ]
+    const handleTenantDeletion = (_res: AxiosResponse<string, ValidationError>, variables: DeleteFnParams) => {
+        const {id} = variables
 
-    const handleTenantDeletion = (id: string) => {
-        console.log(id)
+        setTenants(prevState => {
+            return prevState.filter(tenant => tenant.id !== id)
+        })
     }
 
     const handleOpenTenantDetails = () => {
-        console.log("open")
+        setCurrentTenant(null)
+        setOpenDetails(true)
+    }
+
+    const handleTenantsUpdate = (res: AxiosResponse<Tenant, ValidationError>) => {
+        setTenants(prevState => {
+
+            const filteredTenants = prevState.filter(tenant => tenant.id !== res.data.id)
+
+            return [...filteredTenants, res.data]
+        })
+
+        setOpenDetails(false)
     }
 
     const handleRowClick = (tenant: Tenant) => {
@@ -87,10 +70,10 @@ export default function TenantsPage() {
 
     return (
         <Box sx={{m: 0, p: 2, width: "100%"}}>
-            <DataTable onRowClick={handleRowClick} data={tenants} title="Tenants Table" deleteFn={deleteTenant}
-                       onDeleteSuccess={handleTenantDeletion} headCells={headCells} onCreateBtnClick={handleOpenTenantDetails}/>
-            {currentTenant &&
-                <TenantDetailsDialog open={openDetails} onClose={() => setOpenDetails(false)} value={currentTenant}/>}
+            <TenantsTable data={tenants} onRowClick={handleRowClick} onCreateBtnClick={handleOpenTenantDetails}
+                          onDeleteSuccess={handleTenantDeletion}/>
+            <TenantDetailsDialog open={openDetails} onClose={() => setOpenDetails(false)} value={currentTenant}
+                                 onSuccess={handleTenantsUpdate}/>
         </Box>
     )
 }
